chore(ops): deploy to Github pages with TravisCI
Github pages supports Jekyll but not Hugo. So cheat using travis.

Travis will trigger a Hugo build and will push the resulting static
files to a seperate gh-pages branch. This branch can be deployed with
Github pages.

The static files are pushed to a seperate gh-pages branch so that
the master branch is not polluted.